% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_clustering_metrics.R
\name{bench_clustering_metrics}
\alias{bench_clustering_metrics}
\title{Benchmarking: Calculate clustering evaluation metrics}
\usage{
bench_clustering_metrics(
  predicted_clusters,
  true_labels,
  Metric = c("Purity", "F1", "Jaccard", "ARI")
)
}
\arguments{
\item{predicted_clusters}{The predicted clusters}

\item{true_labels}{The true cluster labels.}

\item{Metric}{The calculated metric, "Purity","F1","Jaccard", or "ARI".}
}
\value{
A dataset containing the score dataset.
}
\description{
This function calculates Purity/F1/Jaccard Index score for each cell type, in the case of the ARI score, it will provide the user a global score.
}
