% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_structural_metrics.R
\name{bench_structural_metrics}
\alias{bench_structural_metrics}
\title{Benchmarking: Calculate structural-aware metrics}
\usage{
bench_structural_metrics(
  Obj,
  Metric = c("Silhouette", "cLISI"),
  cluster,
  reduction
)
}
\arguments{
\item{Obj}{A Seurat Object}

\item{Metric}{The calculated metric, "Silhouette" or "cLISI".}

\item{cluster}{The analyzing clusters' name}

\item{reduction}{The dimensionality reduction which will be used to calculate the score. It should already exist in the seurat object.}
}
\value{
A dataset containing the score dataset.
}
\description{
This function calculates silhouette/cLISI score for each cell type
}
