% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_feature_importance_per_class.R
\name{ds_feature_importance_per_class}
\alias{ds_feature_importance_per_class}
\title{This function calculates the feature importance of the deep learning model}
\usage{
ds_feature_importance_per_class(X, Y, model, classes, features, max_n = 1000)
}
\arguments{
\item{X}{The data with the feature information}

\item{Y}{The data with the output information}

\item{model}{The deep learning model}

\item{classes}{The classes to be grouped}

\item{features}{The manually selected features whose importance will be calculated. If not provided,
the function will use all the features. You can control the number of used features by using the parameter
\code{max_n}.}

\item{max_n}{Number of features selected from X if \code{features} not provided.}
}
\value{
A dataframe with the importance score of each feature
}
\description{
This function generates a dataframe with the feature importance score of the input features calculated
based on the "Permutation Feature Importance" approach.
}
\examples{
ds_feature_importance_per_class(X=x_test,Y=y_test,model=scOMM_model,classes=c(),max_n=500)

}
