% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_summary_barplot.R
\name{ds_summary_barplot}
\alias{ds_summary_barplot}
\title{Cumulative barplot of cell mapped compositions}
\usage{
ds_summary_barplot(class.fac, obs.fac)
}
\arguments{
\item{class.fac}{A named vector of classified cells.}

\item{obs.fac}{A named vector of clusters.}
}
\value{
A cumulative barplot describing the cell identity compositions of the
observed clusters onto the reference
}
\description{
This function identifies true label groups between reference groups and clusters.
}
\examples{
# TODO
}
