% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_thresholds.R
\name{find_optimal_thresholds}
\alias{find_optimal_thresholds}
\title{This function is used to find the optimal threshold for each cell type}
\usage{
find_optimal_thresholds(dnn_model = model, model.data = out)
}
\arguments{
\item{model_data}{The output of the ds_split_data_dnn function.}

\item{prob}{The predicted probability matrix}
}
\value{
List containing the classification output and the probability matrix of the model.
}
\description{
This function is used to find the optimal threshold for each cell type
}
